% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{cellAreaKm2}
\alias{cellAreaKm2}
\title{Exact area of specific cell in square kilometers.}
\usage{
cellAreaKm2(cell)
}
\arguments{
\item{cell}{vector of H3 cells}
}
\value{
the exact area of specific cell in square kilometers.
}
\description{
Exact area of specific cell in square kilometers.
}
\examples{

cellAreaKm2(cell = c("8cbe63562a54bff","8cbe635631103ff"))

}
