% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{cellAreaM2}
\alias{cellAreaM2}
\title{Exact area of specific cell in square meters.}
\usage{
cellAreaM2(cell)
}
\arguments{
\item{cell}{vector of H3 cells}
}
\value{
the exact area of specific cell in square meters.
}
\description{
Exact area of specific cell in square meters.
}
\examples{

cellAreaM2(cell = c("8cbe63562a54bff","8cbe635631103ff"))

}
