% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{cellAreaRads2}
\alias{cellAreaRads2}
\title{Exact area of specific cell in square radians.}
\usage{
cellAreaRads2(cell)
}
\arguments{
\item{cell}{vector of H3 cells}
}
\value{
the exact area of specific cell in square radians.
}
\description{
Exact area of specific cell in square radians.
}
\examples{

cellAreaRads2(cell = c("8cbe63562a54bff","8cbe635631103ff"))

}
