% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direcetedEdges.R
\name{cellsToDirectedEdge}
\alias{cellsToDirectedEdge}
\title{Cells to directed edge}
\usage{
cellsToDirectedEdge(origin, destination)
}
\arguments{
\item{origin}{vector of origin H3 cell indexes}

\item{destination}{vector of destination H3 cell indexes}
}
\value{
a unidirectional edge H3 index based on the provided origin and destination.
}
\description{
Returns a unidirectional edge H3 index based on the provided origin and destination.
}
\examples{

cellsToDirectedEdge(origin = c("85283471fffffff","85283473fffffff"), destination = c("85283475fffffff","85283477fffffff"))

}
