% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{cellsToMultiPolygon}
\alias{cellsToMultiPolygon}
\title{Cells to GeoPolygons}
\usage{
cellsToMultiPolygon(cellSet, isGeoJson = 0L)
}
\arguments{
\item{cellSet}{list of string vectors containing to be H3 cell indexes}

\item{isGeojson}{if the input isGeoJson (each vertex is (lon, lat)). The default is 0L (each vertex is (lat, lon))}
}
\value{
GeoJSON-like MultiPolygon containing the outlines of the set of input H3 hexagons
}
\description{
Get the outlines of a set of H3 hexagons, returned in GeoJSON-like MultiPolygon.
}
\examples{

cellsToMultiPolygon(cellSet = c(c("872830828ffffff", "87283082effffff")))

}
