% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{cellToBoundary}
\alias{cellToBoundary}
\title{Cell To Boundary}
\usage{
cellToBoundary(cell)
}
\arguments{
\item{cell}{vector of H3 cells}
}
\value{
named list, each element named with the input H3 cell, and containing
a `lat` and `lng` vector
}
\description{
Cell To Boundary
}
\examples{

cellToBoundary(cell = c("8cbe63562a54bff","8cbe635631103ff"))

}
