% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{cellToCenterChild}
\alias{cellToCenterChild}
\title{Cell To Center Child}
\usage{
cellToCenterChild(cell, childRes)
}
\description{
Provies the center child index contained by `cell` at the `childRes` resolution
}
