% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{cellToChildren}
\alias{cellToChildren}
\title{Cell To Children}
\usage{
cellToChildren(cell, childRes)
}
\value{
a named list, where each element is the input `cell`, and the values
of each element are the child H3 cells
}
\description{
Returns all the H3 indexes contained by the input `cell` and the defined child resolution
}
\examples{

cell <- "8cbe63562a54bff"
getResolution(cell = cell)

cellToChildren(cell, childRes = 13L)

}
