% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertexes.R
\name{cellToVertex}
\alias{cellToVertex}
\title{Cell To Vertex}
\usage{
cellToVertex(cell, vertexNum)
}
\arguments{
\item{cell}{vector of H3 cells}

\item{vertexNum}{intger giving the vertex number of the index to return}
}
\description{
Returns the index for the specified cell vertex.
Valid vertex numbers are between 0 and 5 (inclusive) for hexagonal cells,
and 0 and 4 (inclusive) for pentagonal cells.
}
\examples{

cellToVertex(
  cell = c(rep("8cbe63562a54bff", 6))
  , vertexNum = c(0L:5L)
  )

}
