% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertexes.R
\name{cellToVertexes}
\alias{cellToVertexes}
\title{Cell To Vertexes}
\usage{
cellToVertexes(cell)
}
\arguments{
\item{cell}{vector of H3 cells}
}
\value{
list of vectors giving the vertices of each cell.
Each list element corresponds to the cell index given in the \code{cell} argument,
and each element of the vector are the cell vertexes.
}
\description{
Returns the indexes for all vertices of the given \code{cell}. The
Output will have a 0 in the result if the input \code{cell} is a pentagon
}
\examples{

cellToVertexes(cell = c("8cbe63562a54bff", "8cbe635631103ff") )


}
