% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{compactCells}
\alias{compactCells}
\title{Compacts the set cellSet of indexes as best as possible, into the array compactedSet.}
\usage{
compactCells(cellSet)
}
\arguments{
\item{cellSet}{list of character vectors containing to be compacted H3 cell indexes}
}
\value{
a list of character vectors containing the compacted H3 cell indexes
}
\description{
Compacts the set cellSet of indexes as best as possible, into the array compactedSet.
}
\examples{

compactCells(gridDisk(cell = c("8cbe63562a54bff","8cbe635631103ff"), k = c(1L, 2L)))

}
