% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{degsToRads}
\alias{degsToRads}
\title{Degrees To Rads}
\usage{
degsToRads(deg)
}
\arguments{
\item{deg}{vector of degrees}
}
\value{
numeric vector giving the input \code{deg} values as radians
}
\description{
Converts Degrees to Radians
}
\examples{

degsToRads(deg = seq(0, 360, by = 15))


}
