% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direcetedEdges.R
\name{directedEdgeToBoundary}
\alias{directedEdgeToBoundary}
\title{Directed edge To Boundary}
\usage{
directedEdgeToBoundary(edge)
}
\arguments{
\item{edge}{unidirectional edge}
}
\value{
named list, each element named with the input H3 cell, and containing
a `lat` and `lng` vector
}
\description{
Directed edge To Boundary
}
\examples{

directedEdgeToBoundary(edge = c("115283473fffffff","115283477fffffff"))

}
