% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symh3_test.R
\name{directions_cell}
\alias{directions_cell}
\title{Get the h3 cells from the origin to an direction within a certain distance}
\usage{
directions_cell(origin_vector, distance_vector, direction_vector)
}
\arguments{
\item{origin_vector}{vector of origin H3 cell indexes}

\item{distance_vector}{int vector of distance}

\item{direction_vector}{string vector of direction. Choices are `c('I', 'J', 'K', 'IJ', 'IK', 'JK')`}
}
\value{
a list of lists of h3 cells on the path of an origin to distance away on direction
}
\description{
Get the h3 cells from the origin to an direction within a certain distance
}
\examples{
directions_cell(origin_vector = c("8cbe63562a54bff", "8cbe635631103ff")
               , distance_vector = c(5L, 100L)
               , direction_vector = c("I", "JK"))

}
