% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{edgeLengthKm}
\alias{edgeLengthKm}
\title{Get the exact edge length of specific unidirectional edge in kilometers.}
\usage{
edgeLengthKm(edge)
}
\arguments{
\item{edge}{vector of unidirectional H3 edges}
}
\value{
the exact edge length of specific unidirectional edge in kilometers.
}
\description{
Get the exact edge length of specific unidirectional edge in kilometers.
}
\examples{

edgeLengthKm(edge = c("13d2a1672b34ffff","16a2a1072b59ffff"))

}
