% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{edgeLengthM}
\alias{edgeLengthM}
\title{Get the exact edge length of specific unidirectional edge in meters.}
\usage{
edgeLengthM(edge)
}
\arguments{
\item{edge}{vector of unidirectional H3 edges}
}
\value{
the exact edge length of specific unidirectional edge in meters.
}
\description{
Get the exact edge length of specific unidirectional edge in meters.
}
\examples{

edgeLengthM(edge = c("13d2a1672b34ffff","16a2a1072b59ffff"))

}
