% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{edgeLengthRads}
\alias{edgeLengthRads}
\title{Get the exact edge length of specific unidirectional edge in radians.}
\usage{
edgeLengthRads(edge)
}
\arguments{
\item{edge}{vector of unidirectional H3 edges}
}
\value{
the exact edge length of specific unidirectional edge in radians.
}
\description{
Get the exact edge length of specific unidirectional edge in radians.
}
\examples{

edgeLengthRads(edge = c("13d2a1672b34ffff","16a2a1072b59ffff"))

}
