% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{getBaseCellNumber}
\alias{getBaseCellNumber}
\title{Get Base Cell Number}
\usage{
getBaseCellNumber(cell)
}
\arguments{
\item{cell}{vector of H3 cells}
}
\value{
a vector the same length as \code{cell} giving the base cell number of
each index
}
\description{
Returns the base cell number of the index
}
\examples{

getBaseCellNumber(cell = c("8cbe63562a54bff","8cbe635631103ff"))

}
