% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBoundingHexFromCoords}
\alias{getBoundingHexFromCoords}
\title{getBoundingHexFromCoords returns the h3 bounding hex from a lat, lon and resolution}
\usage{
getBoundingHexFromCoords(lat, lon, resolution)
}
\arguments{
\item{lat}{latitude in degrees}

\item{lon}{longitude in degrees}

\item{resolution}{resolution, higher is more granular}
}
\description{
getBoundingHexFromCoords returns the h3 bounding hex from a lat, lon and resolution
}
