% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBoundingHexFromIndex}
\alias{getBoundingHexFromIndex}
\title{getBoundingHexFromIndex takes an h3 index and returns the corresponding bounding hex}
\usage{
getBoundingHexFromIndex(index)
}
\arguments{
\item{index}{h3 index as a hex representation character vector. See \code{\link{getIndexFromCoords}}}
}
\description{
getBoundingHexFromIndex takes an h3 index and returns the corresponding bounding hex
}
