% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCIndexFromCoords}
\alias{getCIndexFromCoords}
\title{getCIndexFromCoords returns the raw h3 index from a lat, lon, and resolution}
\usage{
getCIndexFromCoords(lat, lon, res)
}
\arguments{
\item{lat}{latitude in degrees}

\item{lon}{longitude in degrees}

\item{resolution}{resolution, higher is more granular}
}
\description{
getCIndexFromCoords returns the raw h3 index from a lat, lon, and resolution
}
