% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direcetedEdges.R
\name{getDirectedEdgeDestination}
\alias{getDirectedEdgeDestination}
\title{Get Directed Edge Destination}
\usage{
getDirectedEdgeDestination(edge)
}
\arguments{
\item{edge}{vector of unidirectional edge H3 indexes}
}
\value{
the destination hexagon from the unidirectional edge H3Index.
}
\description{
Get the destination cell of the unidirectional edge
}
\examples{

getDirectedEdgeDestination(edge = c("115283473fffffff","16a2a1072b59ffff"))

}
