% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3.R
\name{getHex}
\alias{getHex}
\title{getHex gets the h3 hexagon at the index corresponding to the given coordinates and the given resolution}
\usage{
getHex(lat = 29.7632836, lon = -95.3632715, resolution = 5)
}
\arguments{
\item{lat}{latitude in degrees}

\item{lon}{longitude in degrees}

\item{resolution}{resolution, higher is more granular}
}
\description{
getHex gets the h3 hexagon at the index corresponding to the given coordinates and the given resolution
}
