% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getHexagonAreaAvgKm2}
\alias{getHexagonAreaAvgKm2}
\title{Get the average hexagon area in square kilometers at the given resolution. Excludes pentagons.}
\usage{
getHexagonAreaAvgKm2(resolution)
}
\arguments{
\item{resolution}{cell resolution}
}
\value{
Average hexagon area in square kilometers at the given resolution. Excludes pentagons.
}
\description{
Get the average hexagon area in square kilometers at the given resolution. Excludes pentagons.
}
\examples{

getHexagonAreaAvgKm2(resolution = c(12L,10L))

}
