% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getHexagonEdgeLengthAvgKm}
\alias{getHexagonEdgeLengthAvgKm}
\title{Get the average hexagon edge length in kilometers at the given resolution. Excludes pentagons.}
\usage{
getHexagonEdgeLengthAvgKm(resolution)
}
\arguments{
\item{resolution}{cell resolution}
}
\value{
Average hexagon edge length in kilometers at the given resolution. Excludes pentagons.
}
\description{
Get the average hexagon edge length in kilometers at the given resolution. Excludes pentagons.
}
\examples{

getHexagonEdgeLengthAvgKm(resolution = c(12L,10L))

}
