% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getHexagonEdgeLengthAvgM}
\alias{getHexagonEdgeLengthAvgM}
\title{Get the average hexagon edge length in meters at the given resolution.
Excludes pentagons.}
\usage{
getHexagonEdgeLengthAvgM(resolution)
}
\arguments{
\item{resolution}{cell resolution}
}
\value{
Average hexagon edge length in meters at the given resolution.
Excludes pentagons.
}
\description{
Get the average hexagon edge length in meters at the given resolution.
Excludes pentagons.
}
\examples{

getHexagonEdgeLengthAvgM(resolution = c(12L,10L))

}
