% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getIndexFromCoords}
\alias{getIndexFromCoords}
\title{getIndexFromCoords returns the h3 index from a lat, lon, and resolution}
\usage{
getIndexFromCoords(lat, lon, resolution)
}
\arguments{
\item{lat}{latitude in degrees}

\item{lon}{longitude in degrees}

\item{resolution}{resolution, higher is more granular}
}
\description{
getIndexFromCoords returns the h3 index from a lat, lon, and resolution
}
