% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getNumCells}
\alias{getNumCells}
\title{Get the number of unique H3 indexes at the given resolution.}
\usage{
getNumCells(resolution)
}
\arguments{
\item{resolution}{cell resolution}
}
\value{
the number of unique H3 indexes at the given resolution, in string.
}
\description{
Get the number of unique H3 indexes at the given resolution.
}
\examples{

getNumCells(resolution = c(12L,10L))

}
