% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getPentagons}
\alias{getPentagons}
\title{Get all the pentagon H3 indexes at the specified resolution.}
\usage{
getPentagons(resolution)
}
\arguments{
\item{resolution}{cell resolution}
}
\value{
all the pentagon H3 indexes at the specified resolution.
}
\description{
Get all the pentagon H3 indexes at the specified resolution.
}
\examples{

getPentagons(resolution = c(12L,10L))

}
