% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getRes0Cells}
\alias{getRes0Cells}
\title{Get all the resolution 0 H3 indexes.}
\usage{
getRes0Cells()
}
\value{
all the resolution 0 H3 indexes.
}
\description{
Get all the resolution 0 H3 indexes.
}
\examples{

getRes0Cells()

}
