% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{greatCircleDistanceKm}
\alias{greatCircleDistanceKm}
\title{Great Circle Distance In Kilometers}
\usage{
greatCircleDistanceKm(aLats, aLons, bLats, bLons)
}
\description{
Great Circle Distance In Kilometers
}
\examples{

greatCircleDistanceKm(
  aLats = c(-37.820197)
  , aLons = c(144.983324)
  , bLats = c(-37.818476)
  , bLons = c(144.967354)
)

}
