% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{greatCircleDistanceRads}
\alias{greatCircleDistanceRads}
\title{Great Circle Distance In Radians}
\usage{
greatCircleDistanceRads(aLats, aLons, bLats, bLons)
}
\arguments{
\item{aLats}{vector of latitude cooridnates (from)}

\item{aLons}{vector of longitude coordinates (from)}

\item{bLats}{vector of latitude coordinates (to)}

\item{bLons}{vector of longitude coordinates (to)}
}
\description{
Gives the "great circle" or "haversine" distance between pairs of lat/lng coordinates
in radians
}
\examples{

greatCircleDistanceRads(
  aLats = c(-37.820197)
  , aLons = c(144.983324)
  , bLats = c(-37.818476)
  , bLons = c(144.967354)
)

}
