% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{greatCirlceDistanceRads}
\alias{greatCirlceDistanceRads}
\title{Great Circle Distance In Radians}
\usage{
greatCirlceDistanceRads(aLats, aLons, bLats, bLons)
}
\description{
Great Circle Distance In Radians
}
\examples{

greatCirlceDistanceRads(
  aLats = c(-37.820197)
  , aLons = c(144.983324)
  , bLats = c(-37.818476)
  , bLons = c(144.967354)
)

}
