% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traversal.R
\name{gridDiskDistances}
\alias{gridDiskDistances}
\title{#' Get the maximum number of indices that result from the gridDisk algorithm
#' with the given k.
#'
#' k-ring 0 is defined as the origin index, k-ring 1 is defined as k-ring 0
#' and all neighboring indexes, and so on.
#'
#' @param k int distance
#'
#' @return maximum number of indices that result from the gridDisk algorithm
#' with the given k.
#'
#' @examples
#'
#' maxGridDiskSize(k = c(1L, 2L))
#'
#' @export
maxGridDiskSize <- function(k) {
  .Call(h3rMaxGridDiskSize, k)
}
Grid Disk Distances}
\usage{
gridDiskDistances(cell, k)
}
\arguments{
\item{cell}{vector of H3 cells}

\item{k}{int distance}
}
\value{
indices within k distance of the origin index.
}
\description{
Get indices within k distance of the origin index.
}
\details{
k-ring 0 is defined as the origin index, k-ring 1 is defined as k-ring 0
and all neighboring indexes, and so on.
}
\examples{

gridDiskDistances(cell = c("8cbe63562a54bff","8cbe635631103ff")
                 , k = c(1L, 2L))

}
