% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traversal.R
\name{gridDiskUnsafe}
\alias{gridDiskUnsafe}
\title{Get indices within k distance of the origin index (unsafe version).}
\usage{
gridDiskUnsafe(cell, k)
}
\arguments{
\item{cell}{vector of H3 cells}

\item{k}{int distance}
}
\value{
the indices within k distance of the origin index
}
\description{
k-ring 0 is defined as the origin index, k-ring 1 is defined as k-ring 0
and all neighboring indexes, and so on.
}
\examples{

gridDiskUnsafe(cell = c("8cbe63562a54bff","8cbe635631103ff"), k = c(1L, 2L))

}
