% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traversal.R
\name{gridPathCells}
\alias{gridPathCells}
\title{Grid Path Cells}
\usage{
gridPathCells(origin, destination)
}
\arguments{
\item{origin}{vector of origin H3 cell indexes}

\item{destination}{vector of destination H3 cell indexes}
}
\value{
the line of indexes between the two H3 cells
}
\description{
Given two H3 indexes, return the line of indexes between them (inclusive).
}
\details{
This function may fail to find the line between two indexes, for example if
they are very far apart.
It may also fail when finding distances for indexes on opposite sides of a
pentagon.
}
\examples{

gridPathCells(origin = c("85283473fffffff","85283473fffffff")
              , destination = c("85283471fffffff","85283477fffffff"))

}
