% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traversal.R
\name{gridPathCellsSize}
\alias{gridPathCellsSize}
\title{Get the size of the grid of cells forming the shortest path between two H3 cells.}
\usage{
gridPathCellsSize(origin, destination)
}
\arguments{
\item{origin}{vector of origin H3 cell indexes}

\item{destination}{vector of destination H3 cell indexes}
}
\value{
the size of the grid of cells forming the shortest path between the two H3 cells
}
\description{
Get the size of the grid of cells forming the shortest path between two H3 cells.
}
\examples{

gridPathCellsSize(origin = c("85283473fffffff","85283473fffffff"), destination = c("85283471fffffff","85283477fffffff"))

}
