% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traversal.R
\name{gridRingUnsafe}
\alias{gridRingUnsafe}
\title{Produces the hollow hexagonal ring centered at origin with sides of length k.}
\usage{
gridRingUnsafe(cell, k)
}
\arguments{
\item{cell}{vector of H3 cells}

\item{k}{side length}
}
\value{
the indices of the hollow hexagonal ring centered at origin with sides of length k.
}
\description{
Produces the hollow hexagonal ring centered at origin with sides of length k.
}
\examples{

gridRingUnsafe(cell = c("8cbe63562a54bff","85283473fffffff"), k = c(2L, 1L))

}
