% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-wk.R
\name{as_xy.h3_index}
\alias{as_xy.h3_index}
\alias{as_wkb.h3_index}
\alias{as_wkt.h3_index}
\alias{as_xy.h3_set}
\alias{as_wkb.h3_set}
\alias{as_wkt.h3_set}
\title{Export H3 Index/Set objects to geometry}
\usage{
\method{as_xy}{h3_index}(x, ...)

\method{as_wkb}{h3_index}(x, what, ...)

\method{as_wkt}{h3_index}(x, what, ...)

\method{as_xy}{h3_set}(x, ...)

\method{as_wkb}{h3_set}(x, what, ...)

\method{as_wkt}{h3_set}(x, what, ...)
}
\arguments{
\item{x}{An \code{\link[=h3_index]{h3_index()}} or \code{\link[=h3_set]{h3_set()}}}

\item{...}{Unused}

\item{what}{One of "center", "boundary", or "polygon".
For \code{\link[=h3_index]{h3_index()}} methods, this will return a point,
linestring, or polygon feature. For \code{\link[=h3_set]{h3_set()}} methods,
this will return a multipoint (centre of cells in the set),
a multilinestring (cumulative boundary of cells not including
internal boundaries), or multipolygon (cumulative area of cells
not including internal boundaries).}
}
\description{
Export H3 Index/Set objects to geometry
}
