% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-wk.R
\name{as_h3_index.wk_xy}
\alias{as_h3_index.wk_xy}
\title{Import H3 Index/Set objects from geometry}
\usage{
\method{as_h3_index}{wk_xy}(x, ..., res)
}
\arguments{
\item{x}{A foreign object}

\item{...}{Unused}

\item{res}{An index resolution between 0 (large hexagons)
and 15 (small hexagons).}
}
\value{
An \code{\link[=h3_index]{h3_index()}} or \code{\link[=h3_set]{h3_set()}}
}
\description{
Import H3 Index/Set objects from geometry
}
