% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3-index.R
\name{h3_index}
\alias{h3_index}
\title{Create H3 Index vectors}
\usage{
h3_index(x)
}
\arguments{
\item{x}{The canonical H3 identifier as a character vector.}
}
\value{
A vctr of class h3_index
}
\description{
Create H3 Index vectors
}
\examples{
h3_index("87754e64dffffff")

}
