% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-wk.R
\name{h3_index_point_writer}
\alias{h3_index_point_writer}
\alias{h3_set_writer}
\title{wk package compatibility}
\usage{
h3_index_point_writer(res)

h3_set_writer(res)
}
\arguments{
\item{res}{An index resolution between 0 (large hexagons)
and 15 (small hexagons).}
}
\value{
\itemize{
\item \code{wk_index_point_writer()}: A \link[wk:wk_handle]{wk handler}
}
}
\description{
wk package compatibility
}
\examples{
wk::wk_handle(wk::xy(0, 0), h3_index_point_writer(7))

}
