% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3-set.R
\name{h3_set}
\alias{h3_set}
\title{Create H3 Set vectors}
\usage{
h3_set(h, group_id = 1L)
}
\arguments{
\item{h}{An \code{\link[=h3_index]{h3_index()}} vector.}

\item{group_id}{A vector defining the groups into which
\code{h} should be split. Changes in sequential values define
groups.}
}
\value{
A vctr of class "h3_set"
}
\description{
Create H3 Set vectors
}
