% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3-version.R
\name{h3_version}
\alias{h3_version}
\title{Get the 'H3' library version}
\usage{
h3_version()
}
\value{
A \code{\link[=package_version]{package_version()}} representing the underlying
'H3' C library version.
}
\description{
Get the 'H3' library version
}
\examples{
h3_version()

}
