% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{isPentagon}
\alias{isPentagon}
\title{Is Pentagon}
\usage{
isPentagon(cell)
}
\arguments{
\item{cell}{vector of H3 cells}
}
\value{
a vector the same length as \code{cell} indicating if the cell is Class II
}
\description{
Returns non-zero if this index represents a pentagonal cell.
}
\examples{

isPentagon(cell = c("8cbe63562a54bff","8cbe635631103ff"))

}
