% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{isResClassIII}
\alias{isResClassIII}
\title{Is Res Class III}
\usage{
isResClassIII(cell)
}
\arguments{
\item{cell}{vector of H3 cells}
}
\value{
a vector the same length as \code{cell} indicating if the cell is Class II
}
\description{
Returns non-zero if this index has a resolution with Class III orientation.
}
\examples{

isResClassIII(cell = c("8cbe63562a54bff","8cbe635631103ff"))

hex8 <- "88bf4ac0cdfffff"
hex9 <- "89bf4ac0cd7ffff"
hex10 <- "8abf4ac0cd67fff"
isResClassIII(cell = c(hex8, hex9, hex10))

}
