% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{isValidCell}
\alias{isValidCell}
\title{Is Valid Cell}
\usage{
isValidCell(cell)
}
\arguments{
\item{cell}{vector of H3 cells}
}
\value{
a vector the same length as \code{cell} giving the validity of the cell
}
\description{
Returns non-zero if this is a valid H3 cell index
}
\examples{

isValidCell(cell = c("8cbe63562a54bff","8cbe635631103ff", "abc", "3"))

}
