% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direcetedEdges.R
\name{isValidDirectedEdge}
\alias{isValidDirectedEdge}
\title{Is valid directed edge}
\usage{
isValidDirectedEdge(edge)
}
\arguments{
\item{edge}{vector of unidirectional edge H3 indexes}
}
\value{
1 if it is a unidirectional edge H3Index, otherwise 0.
}
\description{
Determines if the provided H3Index is a valid unidirectional edge index.
}
\examples{

isValidDirectedEdge(edge = c("13d2a1672b34ffff","16a2a1072b59ffff"))

}
