% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertexes.R
\name{isValidVertex}
\alias{isValidVertex}
\title{Is Valid Vertex}
\usage{
isValidVertex(vertex)
}
\arguments{
\item{vertex}{H3 Vertex index}
}
\description{
Returns 1 if the given index represents a valid H3 vertex
}
\examples{

isValidVertex(vertex = c("24cbe63562a549ff", "abc"))


}
