% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{latLngToCell}
\alias{latLngToCell}
\title{lat lng to cell}
\usage{
latLngToCell(lat, lng, resolution)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}

\item{resolution}{cell resolution}
}
\description{
Indexes the location at the specified resolution, returning the index of the
cell containing the location. This buckets the geographic point into the
H3 grid
}
\examples{

latLngToCell(
  lat = c(-37.820197, -37.818476)
  , lng = c(144.983324, 144.967354)
  , resolution = 12L
  )

}
