% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traversal.R
\name{localIjToCell}
\alias{localIjToCell}
\title{Local IJ To Cell}
\usage{
localIjToCell(origin, i, j)
}
\arguments{
\item{origin}{vector of anchor cell}

\item{i}{vector of local I coordinate}

\item{j}{vector of local I coordinate}
}
\value{
cell vector of H3 cells
}
\description{
Produces an H3 index from local IJ coordinates anchored by an origin.
}
\examples{

localIjToCell(origin = c("85283473fffffff","85283473fffffff")
             , i = c(1L, 2L), j = c(2L, 1L))

}
