% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{maxFaceCount}
\alias{maxFaceCount}
\title{Max Face Count}
\usage{
maxFaceCount(cell)
}
\description{
Returns the maximum number of icosahedron faces the given H3 index may intersect.
}
\examples{
maxFaceCount(cell = c("8cbe63562a54bff","8cbe635631103ff"))


}
