% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traversal.R
\name{maxGridDiskSize}
\alias{maxGridDiskSize}
\title{Get the maximum number of indices that result from the gridDisk algorithm with the given k.}
\usage{
maxGridDiskSize(k)
}
\arguments{
\item{k}{int distance}
}
\value{
maximum number of indices that result from the gridDisk algorithm with the given k.
}
\description{
k-ring 0 is defined as the origin index, k-ring 1 is defined as k-ring 0
and all neighboring indexes, and so on.
}
\examples{

maxGridDiskSize(k = c(1L, 2L))

}
