% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{pentagonCount}
\alias{pentagonCount}
\title{Get the number of pentagon H3 indexes per resolution. This is always 12, but provided as a convenience.}
\usage{
pentagonCount()
}
\value{
the number of pentagon H3 indexes per resolution.
}
\description{
Get the number of pentagon H3 indexes per resolution. This is always 12, but provided as a convenience.
}
\examples{

pentagonCount()

}
