% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{res0CellCount}
\alias{res0CellCount}
\title{Get the number of resolution 0 H3 indexes, which is defined as 122.}
\usage{
res0CellCount()
}
\value{
the number of resolution 0 H3 indexes, which is defined as 122.
}
\description{
Get the number of resolution 0 H3 indexes, which is defined as 122.
}
\examples{

res0CellCount()

}
