% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3r-package.R
\docType{data}
\name{stations}
\alias{stations}
\title{Stations}
\format{
\describe{
\item{stop_id}{The ID of the station}
\item{stop_name}{The name of the station}
\item{lat}{The latitude coordinate of the station}
\item{lon}{The longitude coordinate of the station}
}
}
\usage{
stations
}
\description{
A data set of train stations in Melbourne
}
\details{
Obtained from \url{https://discover.data.vic.gov.au/dataset/gtfs-schedule}
and distributed under the Creative Commons 4 License \url{https://creativecommons.org/licenses/by/4.0/}
}
\keyword{datasets}
