% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk-writer.R
\name{wk_writer}
\alias{wk_writer}
\alias{h3_cell_writer}
\alias{listof_h3_cell_writer}
\title{WK Writers}
\usage{
h3_cell_writer(res)

listof_h3_cell_writer(res)
}
\arguments{
\item{res}{An index resolution between 0 (large hexagons)
and 15 (small hexagons).}
}
\value{
\itemize{
\item \code{h3_cell_writer()}: A \link[wk:wk_handle]{wk handler}
\item \code{listof_h3_cell_writer()}: A \link[wk:wk_handle]{wk handler}
}
}
\description{
WK Writers
}
\examples{
wk::wk_handle(wk::xy(0, 0), h3_cell_writer(7))
wk::wk_handle(wk::wkt("MULTIPOINT ((0 0, 1 1))"), listof_h3_cell_writer(7))

}
