% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direcetedEdges.R
\name{getDirectedEdgeOrigin}
\alias{getDirectedEdgeOrigin}
\title{Get Directed Edge Origin}
\usage{
getDirectedEdgeOrigin(edge)
}
\arguments{
\item{edge}{vector of unidirectional edge H3 indexes}
}
\value{
the origin hexagon from the unidirectional edge H3Index.
}
\description{
Get the origin cell of the unidirectional edge
}
\examples{

getDirectedEdgeOrigin((edge = c("115283473fffffff","16a2a1072b59ffff")))

}
