% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{greatCircleDistanceKm}
\alias{greatCircleDistanceKm}
\title{Great Circle Distance In Kilometers}
\usage{
greatCircleDistanceKm(aLat, aLng, bLat, bLng)
}
\arguments{
\item{aLat}{vector of latitude cooridnates (from)}

\item{aLng}{vector of longitude coordinates (from)}

\item{bLat}{vector of latitude coordinates (to)}

\item{bLng}{vector of longitude coordinates (to)}
}
\description{
Gives the "great circle" or "haversine" distance between pairs of lat/lng coordinates
in kilometers.
}
\examples{

greatCircleDistanceKm(
  aLat = c(-37.820197)
  , aLng = c(144.983324)
  , bLat = c(-37.818476)
  , bLng = c(144.967354)
)

}
