% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traversal.R
\name{gridDistance}
\alias{gridDistance}
\title{Grid Distance}
\usage{
gridDistance(origin, destination)
}
\arguments{
\item{origin}{vector of origin H3 cell indexes}

\item{destination}{vector of destination H3 cell indexes}
}
\value{
the grid distance between the two H3 cells
}
\description{
Provides the distance in grid cells between the two indexes.
}
\examples{

gridDistance(origin = c("85283473fffffff","85283473fffffff"),
             destination = c("8528342bfffffff","85283477fffffff"))

}
