% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3r-package.R
\docType{package}
\name{h3r-package}
\alias{h3r}
\alias{h3r-package}
\title{h3r: Bindings for the 'H3' Index System}
\description{
Provides bindings for the 'H3' heagonal index library.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
