% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{polygonToCells}
\alias{polygonToCells}
\title{Polygon to cells}
\usage{
polygonToCells(polygons, resolution, isGeoJson = FALSE)
}
\arguments{
\item{polygons}{an array of polygons. Each polygon is an array of vectors. Each vector is a list of (lat, lon) or (lon, lat)}

\item{resolution}{The resolution of the output cells}

\item{isGeojson}{if the input isGeoJson (each vertex is (lon, lat)). The default is FALSE (each vertex is (lat, lon))}
}
\description{
Returns the h3 indexes for the input GeoJSON-like data structure
}
\examples{

cellToVertex(
  cell = c(rep("8cbe63562a54bff", 6))
  , vertexNum = c(0L:5L)
  )

}
