% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{radsToDegs}
\alias{radsToDegs}
\title{Rads to Degrees}
\usage{
radsToDegs(rad)
}
\arguments{
\item{rad}{vector of radians}
}
\value{
numeric vector giving the input \code{rad} values as degrees
}
\description{
Converts Radians to Degrees
}
\examples{
radsToDegs(rad = seq(0, 2 * pi, by = (pi / 12) ) )

}
