% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertexes.R
\name{vertexToLatLng}
\alias{vertexToLatLng}
\title{Vertex To Lat Lng}
\usage{
vertexToLatLng(vertex)
}
\arguments{
\item{vertex}{H3 Vertex index}
}
\description{
Returns the latitude and longitude of the given vertex
}
\examples{

vertexToLatLng(
  vertex = cellToVertex(
    cell = c("8cbe63562a54bff", "8cbe635631103ff")
    , vertexNum = c(1L, 3L)
    )
)


}
