% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertexes.R
\name{vertexToLatLng}
\alias{vertexToLatLng}
\title{Vertex To Lat Lng}
\usage{
vertexToLatLng(vertex)
}
\arguments{
\item{vertex}{H3 Vertex index}
}
\value{
\code{data.frame} of the lat/lng coordinates of the input \code{vertex}
}
\description{
Returns the latitude and longitude of the given vertex
}
\examples{

vertices <- cellToVertex(
  cell = c(rep("8cbe63562a54bff", 6))
  , vertexNum = c(0L:5L)
  )

vertexToLatLng(vertex = vertices)


}
